backend/scripts/requirements.txt
yfinance
pandas
pymongo
python-dotenv